requirements.txt
flask
fpdf
pillow
pytesseract
python-dotenv
PyMuPDF
google-generativeai